% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.LP.R
\name{as.LP}
\alias{as.LP}
\title{Convert a data frame to a \code{LP} problem}
\usage{
as.LP(x, integer = NULL, binary = NULL)
}
\arguments{
\item{x}{A linear programming problem in tabular form.}
}
\description{
Convert a data frame to a \code{LP} problem
}
