% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.PMP.R
\name{as.PMP}
\alias{as.PMP}
\title{Convert an \code{LP} object to a \code{PMP} object}
\usage{
as.PMP(LP, xstar, Q = NULL)
}
\arguments{
\item{LP}{A linear programming problem as an \code{LP} object}

\item{xstar}{Observations of \code{x}}

\item{Q}{The Q-matrix for the PMP problem}
}
\description{
Convert an \code{LP} object to a \code{PMP} object
}
