% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.src.R
\name{barplot.src}
\alias{barplot.src}
\title{Plot Standardized Regression Coefficients}
\usage{
\method{barplot}{src}(
  x,
  ylim = NULL,
  col = NA,
  las = 1,
  names.arg = colnames(x$X),
  main = NA,
  xlab = "",
  ylab = "SRC",
  border = TRUE,
  ...
)
}
\arguments{
\item{x}{A "src" class object containing information about the standardized regression coefficients}

\item{ylim}{Vector containing the lower and upper limit of the y-axis.}

\item{col}{Bar fill color.}

\item{las}{Label style, see ?par() for further information.}

\item{names.arg}{Names below the bars.}

\item{main}{Cutom main title.}

\item{border}{Should a box be drawn around the plot?}
}
\description{
This function generates a barplot of the "Standardized Regression Coefficients" in an \code{src} object.
The arrows on the bars indicate the upper and lower confidence intervals, 95\% by default or any others which were manually set.
}
