% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\title{Function to combine constraints and limits}
\usage{
combine(constraints, limits, objective)
}
\arguments{
\item{constraints}{A data frame containing constraints in the usual format}

\item{limits}{A data frame containing the limits as in the template}
}
\description{
Function to combine constraints and limits
}
