% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duals.R
\name{duals}
\alias{duals}
\title{Shadow prices}
\usage{
duals(LP, what = NA)
}
\arguments{
\item{LP}{An "LP" class object}

\item{what}{Name of a constraint; if selected, the dual value of only this constrant will be returned as a single value.}
}
\description{
This function returns shadow prices (a.k.a. the duals) of an \code{LP} object.
}
