% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LP.R
\name{LP}
\alias{LP}
\title{Create an \code{LP} object for linear programming}
\usage{
`LP(c, A, b, ...)`
}
\arguments{
\item{c}{A vector containing the target function coefficients.}

\item{A}{A matrix with the technical coefficients.}

\item{b}{A vector containing the ressource endowments.}

\item{maximize}{Direction of the optimisation task. If \code{TRUE}, the problem gets maximized, if \code{FALSE}, it gets minimized.}

\item{integer}{Vector of indices, which variables should only allowed to be integer?}

\item{binary}{Vector of indices, which variables should only allowed to be binary?}

\item{direction}{Direction of the constraints. By default, these are all "<=" (smaller than or equal to). However, it could also be ">=", "<", ">", or "=". Alternative directions need to be provided as a vector with the same length as \code{b}.}
}
\description{
This function is designed to set up an \code{LP}-class object, which offers a range of further functions to work with.
}
\note{
If the matrix A already has row- or column names, those are prioritised over element names of vectors c and b.
}
