% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{objective}
\alias{objective}
\title{Get the objective value from an \code{PMP} object}
\usage{
objective(x)
}
\arguments{
\item{x}{An "LP"-class object whose objective value should be retrieved.}
}
\description{
Get the objective value from an \code{PMP} object
}
