% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Duals.R
\name{plot.Duals}
\alias{plot.Duals}
\title{Generate a plot of the duals}
\usage{
\method{plot}{Duals}(x)
}
\arguments{
\item{x}{A object of class \code{Duals} with computed shadow prices from an \code{LP} object.}
}
\description{
This function generates a dotchart of the duals (the shadow prices) from an \code{LP} object of class \code{Duals}.
For the function to work properly, the duals must have been calculated prior to the function call.
}
