% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.LP.R
\name{plot.LP}
\alias{plot.LP}
\title{Plot a linear programming problem}
\usage{
\method{plot}{LP}(LP, which = 1:2, col = NULL, lty = NULL, lwd = 1)
}
\arguments{
\item{LP}{A "LP" class object which we wish to plot.}

\item{which}{A vector with exactly two indices of which variables should be plotted.}
}
\description{
This function can plot an "LP" object with exactly two variables.
If there are more than two variables, the function will plot only the first two, or any specified set of variables.
The legend will either use added constraint names or letters, beginning from A.
}
