% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Sensitivity.R
\name{plot.Sensitivity}
\alias{plot.Sensitivity}
\title{Generate a plot of the variable sensitivities}
\usage{
\method{plot}{Sensitivity}(x)
}
\arguments{
\item{x}{A object of class \code{Sensitivity} with computed sensitivities from an \code{LP} object.}
}
\description{
This function generates a dotchart of the sensitivities from an \code{LP} object of class \code{Sensitivity}.
For the function to work properly, the sensitivities must have been calculated prior to the function call.
}
