% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simdata.R
\name{plot.simdata}
\alias{plot.simdata}
\title{Plot Simulated Data for a \code{LP} Sensitivity Analysis}
\usage{
\method{plot}{simdata}(simdata, y = "Objective")
}
\arguments{
\item{simdata}{A \code{simdata} class object to be inspected.}

\item{y}{Name of the dependent variable. By default, this is the value of the objective function ("Objective"). Can be any other column name in \code{simdata$Output}.}
}
\description{
This function generates scatterplot of the data simulated by the \code{simulate.LP()} function.
}
