% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMP.R
\name{PMP}
\alias{PMP}
\title{Create an \code{PMP} object for positive mathematical programming}
\usage{
`PMP(c, A, b, xstar, Q, ...)`
}
\arguments{
\item{c}{A vector containing the target function coefficients.}

\item{A}{A matrix with the technical coefficients.}

\item{b}{A vector containing the ressource endowments.}

\item{Q}{Q-matrix with the elaticities. If not specified, the Q-matrix is directly estimated.}

\item{maximize}{Direction of the optimisation task. If \code{TRUE}, the problem gets maximized, if \code{FALSE}, it gets minimized.}
}
\description{
This function is designed to set up an \code{PMP}-class object, which offers a range of further functions to work with.
}
