% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.PMP.R
\name{predict.PMP}
\alias{predict.PMP}
\title{Create an \code{PMP} object for positive mathematical programming}
\usage{
`PMP(c, A, b, xstar, Q, ...)`
}
\arguments{
\item{PMP}{An already specified positive mathematical programming model of class \code{PMP}}

\item{c}{A vector containing the new target function coefficients with which we want to run the model.}
}
\description{
This function is designed to set up an \code{PMP}-class object, which offers a range of further functions to work with.
}
