% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.PMP.R
\name{print.PMP}
\alias{print.PMP}
\title{Generic print function for \code{PMP} objects}
\usage{
\method{print}{PMP}(PMP)
}
\arguments{
\item{PMP}{}
}
\description{
Generic print function for \code{PMP} objects
}
