% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho.R
\name{rho}
\alias{rho}
\title{Duals of the target function coefficients}
\usage{
rho(LP, what = NULL, bounds = FALSE)
}
\arguments{
\item{LP}{An \code{LP} class object}

\item{what}{Name or index a variable; if selected, the dual value of only this variable will be returned as a single value.}

\item{bounds}{Should the upper and lower bounds of the duals be returned?}
}
\description{
This function returns the duals referring to coefficients (not constraints) of an \code{LP} object.
}
