% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivity}
\alias{sensitivity}
\title{Target function sensitivity}
\usage{
sensitivity(LP)
}
\arguments{
\item{LP}{An "lp" class object with computed sensitivities.}
}
\description{
This function returns the sensitivities of the target function coefficients from an \code{LP} object.
}
