% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.LP.R
\name{simulate.LP}
\alias{simulate.LP}
\title{Simulate multiple "lp" model runs with varying coefficients}
\usage{
\method{simulate}{LP}(LP, n = 1000, range = 0.5)
}
\arguments{
\item{LP}{A "lp" class object.}

\item{n}{Number of simulation runs. Default is n = 1000.}

\item{range}{The range, in which the coefficients of the linear programming problem might vary. Default is range = 0.5, which means the coefficients for the simulations are uniformly distributed in a range of ± 50 \%. It can also be a matrix with upper and lower bounds for the uniform distribution.}
}
\description{
This function generates a dataset with \code{n} simulated observations.
If \code{range} is a single numeric value, the coefficients \eqn{\beta_{\text{sim}}} are sampled from an uniform distribution with:
\deqn{\beta_{\text{sim}} \sim U(\beta^*(1-r), \ \beta^*(1+r))}
where \eqn{\beta^*} are the original coefficients, and \eqn{r} is the specified \code{range} parameter.
Alternatively, a matrix with as many columns as variables can be passed to \code{range}, where the rows contain lower and upper limits for the uniform distributions.
The coefficients are sampled via an improved latin hypercube sampling method in order to achieve a better spread of the sampling points in higher dimensions.
}
