% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.LP.R
\name{update.LP}
\alias{update.LP}
\title{Re-calculate an "LP" object with changed internal states}
\usage{
\method{update}{LP}(
  LP,
  c = LP$c,
  A = LP$A,
  b = LP$b,
  maximize = LP$maximize,
  directions = LP$directions,
  integer = LP$integer,
  binary = LP$binary
)
}
\description{
Re-calculate an "LP" object with changed internal states
}
